backend/pyproject.toml
[tool.poetry]
name = "authly"
version = "0.1.0"
description = "user key managment"
authors = ["wavy42 <wavy_uwu@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
colorama = "^0.4.6"
fastapi = {extras = ["all"], version = "^0.104.1"}
motor = "^3.3.1"
bcrypt = "^4.0.1"
argon2-cffi = "^23.1.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
redis = "^5.0.1"
exrex = "^0.11.0"


[tool.poetry.group.dev.dependencies]
coverage = "^7.3.2"
pytest = "^7.4.3"
snakeviz = "^2.2.0"
httpx = "^0.25.1"
asgi-lifespan = "^2.1.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
bump-pydantic = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
